entity relationship 모델 추가

+++ services/vdap-server/src/main/resources/db/migration/V1__init.sql
@@ -14,7 +14,6 @@ create table storage_service_entity (
 create index storage_service_entity_name_index on storage_service_entity (name);
 create index index_storage_service_entity_deleted on storage_service_entity (id, deleted);
 
-
 create table classification
 (
     id         varchar(36)   primary key,
@@ -39,10 +38,10 @@ create table tag (
 create index tag_entity_name_index on tag (classification_id, name);
 
 create table tag_usage (
-    source      tinyint      not null,
+    source      int          not null,
     tag_id      varchar(256) not null,
-    label_type  tinyint      not null,
-    state       tinyint      not null,
+    label_type  int          not null,
+    state       int          not null,
     target_type varchar(32)  not null,
     target_id   varchar(36)  not null,
     constraint tag_usage_key
@@ -51,14 +50,14 @@ create table tag_usage (
 
 create table entity_relationship
 (
-    from_id     varchar(36)          not null,
-    to_id       varchar(36)          not null,
-    from_entity varchar(256)         not null,
-    to_entity   varchar(256)         not null,
-    relation   tinyint              not null,
-    json_schema varchar(256)         null,
-    json       json                 null,
-    deleted    tinyint(1) default 0 not null,
+    from_id     varchar(36)  not     null,
+    to_id       varchar(36)  not     null,
+    from_entity varchar(256) not     null,
+    to_entity   varchar(256) not     null,
+    relation    int          not     null,
+    json_schema varchar(256) null,
+    json        json         null,
+    deleted     tinyint(1)   default 0     not null,
     primary key (from_id, to_id, relation)
 );
 
@@ -75,6 +74,3 @@ create index idx_er_fromEntity_fromId_toEntity_relation
 create index to_entity_type_index on entity_relationship (to_id, to_entity);
 
 create index to_index on entity_relationship (to_id, relation);
-
-
-
